requirements.txt
APScheduler==3.10.1
boto3==1.26.152
botocore==1.29.152
cachetools==5.3.1
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.4
ConfigArgParse==1.2.3
coverage==7.2.7
cryptography==3.4.8
cytoolz==0.12.1
decorator==5.1.1
Flask==2.3.2
Flask-RESTful==0.3.10
frozendict==2.3.8
frozenlist==1.3.3
future==0.18.2
httplib2==0.20.2
idna==2.10
imagesize==1.4.1
importlib-metadata==4.6.4
indy==0.1
inflection==0.5.1
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
jsonpath-ng==1.5.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.1.1
MarkupSafe==2.1.3
marshmallow==3.5.1
more-itertools==8.10.0
morphys==1.0
msgpack==0.6.2
nats-py==2.3.1
ngrok==0.8.1
oauthlib==3.2.0
Pillow==9.0.1
platformdirs==3.8.0
py-multicodec==0.2.1
pycodestyle==2.6.0
pycparser==2.21
pydantic==1.9.2
pydevd==1.5.1
pydevd-pycharm==193.6015.41
pydid==0.3.8
pydocstyle==3.0.0
pyflakes==2.2.0
PyJWT==1.7.1
PyLD==2.0.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==5.4.3
pytest-asyncio==0.14.0
pytest-cov==2.10.1
pytest-flake8==1.0.6
python-dateutil==2.8.1
pytz==2021.3
pyxdg==0.27
PyYAML==5.3.1
qrcode==6.1
reportlab==3.6.8
requests==2.23.0
rlp==1.2.0
s3transfer==0.6.1
SecretStorage==3.3.1
schedule==1.2.0
six==1.16.0
toolz==0.12.0
tqdm==4.65.0
typing_extensions==4.0.1
unflatten==0.1.1
urllib3==1.25.11
varint==1.0.2
virtualenv==20.23.1
wadllib==1.3.6
wcwidth==0.2.6
webargs==5.5.3
Werkzeug==2.3.6
xdg==5
zipp==1.0.0
